ナレッジベースAPIのコードが2つあったので統合
