+++ app.py
@@ -1,18 +1,28 @@
-from fastapi import FastAPI, HTTPException, Query
+import json
+import os
+import time
+from contextlib import asynccontextmanager
+from typing import Any, Dict, List, Optional
+
+from dotenv import load_dotenv
+from fastapi import FastAPI, File, HTTPException, UploadFile
 from fastapi.responses import JSONResponse
 from pydantic import BaseModel
-from typing import List, Optional, Dict, Any
-import time
-import os
-from contextlib import asynccontextmanager
 
 from embedding_model import PlamoEmbedding
 from opensearch_client import OpenSearchVectorStore
 
+load_dotenv('.env')
 
-class AddDocumentRequest(BaseModel):
+class DocumentRequest(BaseModel):
     content: str
     metadata: Optional[Dict[str, Any]] = {}
+
+
+class SearchRequest(BaseModel):
+    query: str
+    k: Optional[int] = 10
+    filter_query: Optional[Dict] = None
 
 
 class SearchResult(BaseModel):
@@ -23,6 +33,12 @@ class SearchResult(BaseModel):
     timestamp: int
 
 
+class IndexStats(BaseModel):
+    document_count: int
+    size_bytes: int
+    size_mb: float
+
+
 embedding_model = None
 vector_store = None
 INDEX_NAME = "knowledge-base"
@@ -31,35 +47,35 @@ INDEX_NAME = "knowledge-base"
 @asynccontextmanager
 async def lifespan(app: FastAPI):
     global embedding_model, vector_store
-    
+
     print("ナレッジベースAPIを初期化中...")
-    
-    embedding_model = PlamoEmbedding("pfnet/plamo-embedding-1b")
-    
+
+    embedding_model = PlamoEmbedding("./plamo-embedding-1b")
+
     opensearch_host = os.getenv("OPENSEARCH_HOST", "localhost")
     opensearch_port = int(os.getenv("OPENSEARCH_PORT", "9200"))
     opensearch_user = os.getenv("OPENSEARCH_USER", "admin")
-    opensearch_pass = os.getenv("OPENSEARCH_PASS", "admin")
-    
+    opensearch_pass = os.getenv("OPENSEARCH_INITIAL_ADMIN_PASSWORD", "admin")
+
     vector_store = OpenSearchVectorStore(
         host=opensearch_host,
-        port=opensearch_port,  
+        port=opensearch_port,
         username=opensearch_user,
         password=opensearch_pass
     )
-    
+
     embedding_dim = embedding_model.get_embedding_dimension()
     vector_store.create_index(INDEX_NAME, embedding_dim)
-    
+
     print("ナレッジベースAPI初期化完了")
     yield
-    
+
     print("ナレッジベースAPIをシャットダウン中...")
 
 
 app = FastAPI(
     title="ナレッジベースAPI",
-    description="PlamoEmbeddingとOpenSearchを使用した問い合わせ・追加API",
+    description="PlamoEmbeddingとOpenSearchを使用したドキュメント保存・検索API",
     version="1.0.0",
     lifespan=lifespan
 )
@@ -70,39 +86,90 @@ async def root():
     return {"message": "ナレッジベースAPIへようこそ"}
 
 
-@app.get("/search", response_model=List[SearchResult])
-async def search_documents(
-    q: str = Query(..., description="検索クエリ"),
-    k: int = Query(10, description="取得する結果数"),
-    min_score: float = Query(0.0, description="最小スコア閾値")
-):
-    """
-    GET リクエストでドキュメントを検索します
-    
-    - **q**: 検索したいクエリ文字列
-    - **k**: 取得する結果の最大数（デフォルト: 10）
-    - **min_score**: 最小スコア閾値（デフォルト: 0.0）
-    """
+@app.post("/documents")
+async def add_document(request: DocumentRequest):
     try:
         start_time = time.perf_counter()
-        
-        query_embedding = embedding_model.encode([q])
-        
+
+        embedding = embedding_model.encode([request.content])
+
+        document = {
+            "content": request.content,
+            "metadata": request.metadata
+        }
+
+        success_count, failed_items = vector_store.add_documents(
+            INDEX_NAME,
+            [document],
+            embedding
+        )
+
+        end_time = time.perf_counter()
+
+        if failed_items:
+            raise HTTPException(status_code=500, detail="ドキュメントの保存に失敗しました")
+
+        return {
+            "message": "ドキュメントが正常に追加されました",
+            "processing_time": round(end_time - start_time, 2),
+            "embedding_dimension": embedding.shape[1]
+        }
+
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"エラー: {str(e)}")
+
+
+@app.post("/documents/batch")
+async def add_documents_batch(requests: List[DocumentRequest]):
+    try:
+        start_time = time.perf_counter()
+
+        contents = [req.content for req in requests]
+        embeddings = embedding_model.encode(contents)
+
+        documents = [
+            {
+                "content": req.content,
+                "metadata": req.metadata
+            }
+            for req in requests
+        ]
+
+        success_count, failed_items = vector_store.add_documents(
+            INDEX_NAME,
+            documents,
+            embeddings
+        )
+
+        end_time = time.perf_counter()
+
+        return {
+            "message": f"{success_count}件のドキュメントが追加されました",
+            "success_count": success_count,
+            "failed_count": len(failed_items) if failed_items else 0,
+            "processing_time": round(end_time - start_time, 2)
+        }
+
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"エラー: {str(e)}")
+
+
+@app.post("/search", response_model=List[SearchResult])
+async def search_documents(request: SearchRequest):
+    try:
+        start_time = time.perf_counter()
+
+        query_embedding = embedding_model.encode([request.query])
+
         results = vector_store.search(
             INDEX_NAME,
             query_embedding[0],
-            k=k
+            k=request.k,
+            filter_query=request.filter_query
         )
-        
-        # 最小スコア閾値でフィルタリング
-        filtered_results = [
-            result for result in results 
-            if result["score"] >= min_score
-        ]
-        
+
         end_time = time.perf_counter()
-        processing_time = round(end_time - start_time, 2)
-        
+
         search_results = [
             SearchResult(
                 id=result["id"],
@@ -111,78 +178,47 @@ async def search_documents(
                 metadata=result["metadata"],
                 timestamp=result["timestamp"]
             )
-            for result in filtered_results
+            for result in results
         ]
-        
+
         return search_results
-        
+
     except Exception as e:
         raise HTTPException(status_code=500, detail=f"検索エラー: {str(e)}")
 
 
-@app.post("/add")
-async def add_document(request: AddDocumentRequest):
-    """
-    新しいドキュメントを追加します
-    
-    - **content**: 追加するドキュメントの内容
-    - **metadata**: オプションのメタデータ
-    """
-    try:
-        start_time = time.perf_counter()
-        
-        embedding = embedding_model.encode([request.content])
-        
-        document = {
-            "content": request.content,
-            "metadata": request.metadata
-        }
-        
-        success_count, failed_items = vector_store.add_documents(
-            INDEX_NAME, 
-            [document], 
-            embedding
-        )
-        
-        end_time = time.perf_counter()
-        
-        if failed_items:
-            raise HTTPException(status_code=500, detail="ドキュメントの保存に失敗しました")
-        
-        return {
-            "message": "ドキュメントが正常に追加されました",
-            "processing_time": round(end_time - start_time, 2),
-            "embedding_dimension": embedding.shape[1],
-            "document_id": success_count
-        }
-        
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"追加エラー: {str(e)}")
-
-
-@app.get("/stats")
-async def get_stats():
-    """
-    インデックスの統計情報を取得します
-    """
+@app.get("/stats", response_model=IndexStats)
+async def get_index_stats():
     try:
         stats = vector_store.get_index_stats(INDEX_NAME)
-        
+
         if "error" in stats:
             raise HTTPException(status_code=404, detail=stats["error"])
-        
-        return {
-            "index_name": INDEX_NAME,
-            "document_count": stats["document_count"],
-            "size_bytes": stats["size_bytes"],
-            "size_mb": stats["size_mb"]
-        }
-        
+
+        return IndexStats(
+            document_count=stats["document_count"],
+            size_bytes=stats["size_bytes"],
+            size_mb=stats["size_mb"]
+        )
+
     except HTTPException:
         raise
     except Exception as e:
         raise HTTPException(status_code=500, detail=f"統計取得エラー: {str(e)}")
 
+
+@app.delete("/index")
+async def reset_index():
+    try:
+        vector_store.delete_index(INDEX_NAME)
+
+        embedding_dim = embedding_model.get_embedding_dimension()
+        vector_store.create_index(INDEX_NAME, embedding_dim)
+
+        return {"message": "インデックスがリセットされました"}
+
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"インデックスリセットエラー: {str(e)}")
 
 if __name__ == "__main__":
     import uvicorn
